module refactor; rename executable to 'bdb'
