+++ bdb_transaction_cli/__main__.py
@@ -1,0 +1,5 @@
+
+from bdb_transaction_cli.cli import main
+
+if __name__ == '__main__':
+    main()
