+++ bdb_transaction_cli/cli.py
@@ -4,17 +4,7 @@ import json
 from bigchaindb_common.crypto import generate_key_pair
 from bigchaindb_common.transaction import Transaction, Condition
 
-
-class JsonParamType(click.ParamType):
-    """ A Click parameter class that loads JSON """
-    name = 'JSON'
-
-    def convert(self, value, param, ctx):
-        # No try-except in case `value is None`, as it is a valid value.
-        # Still we want `json.loads` to visibly fail.
-        if value is None:
-            return value
-        return json.loads(value)
+from bdb_transaction_cli.utils import JsonParamType
 
 
 @click.group()
@@ -110,7 +100,3 @@ def sign(transaction, private_key):
 
 
 # TODO: bdb transfer '[.. ffills]' '[.. conditions]' '[payload]'
-
-
-if __name__ == '__main__':
-    main()
